backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Validation & Models
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS & Security
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
